Adding some extra logs to track which is FP32

+++ examples/controlnet/train_controlnet_sdxl.py
@@ -936,7 +936,8 @@ def main(args):
 
     if args.controlnet_model_name_or_path:
         logger.info("Loading existing controlnet weights")
-        controlnet = ControlNetModel.from_pretrained(args.controlnet_model_name_or_path)
+        controlnet = ControlNetModel.from_pretrained(args.controlnet_model_name_or_path,weight_dtype = torch.bfloat16 )
+        
     else:
         logger.info("Initializing controlnet weights from unet")
         controlnet = ControlNetModel.from_unet(unet)
